Refactored parallelism utilities and implemented threads option for cortex cli parallel search

+++ cortex-processing/src/main/kotlin/com/hileco/cortex/processing/web3rpc/parallelism/ResumeCounter.kt
@@ -1,20 +1,20 @@
 package com.hileco.cortex.processing.web3rpc.parallelism
 
-class ResumeCounter(initialValue: Int) {
+class ResumeCounter(initialValue: Long) {
     private val monitor = Object()
-    private val idsCompleted = mutableListOf<Int>()
+    private val idsCompleted = mutableListOf<Long>()
     private var resumeFrom = initialValue
 
-    fun value(): Int {
+    fun value(): Long {
         return resumeFrom
     }
 
-    fun complete(index: Int) {
+    fun complete(index: Long) {
         synchronized(monitor) {
             idsCompleted.add(index)
             idsCompleted.removeIf { it < resumeFrom }
             val highestValue = idsCompleted.max() ?: index
-            val nextResumeFrom = IntRange(resumeFrom, highestValue).firstOrNull { !idsCompleted.contains(it) } ?: highestValue + 1
+            val nextResumeFrom = LongRange(resumeFrom, highestValue).firstOrNull { !idsCompleted.contains(it) } ?: highestValue + 1
             resumeFrom = nextResumeFrom
         }
     }
